refactor: create repository package
