+++ pkg/registry/apis/provisioning/files.go
@@ -13,6 +13,7 @@ import (
 	"k8s.io/apiserver/pkg/registry/rest"
 
 	provisioning "github.com/grafana/grafana/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository"
 )
 
 type filesConnector struct {
@@ -110,7 +111,7 @@ func (s *filesConnector) Connect(ctx context.Context, name string, opts runtime.
 	}), nil
 }
 
-func (s *filesConnector) getParser(repo Repository) (*fileParser, error) {
+func (s *filesConnector) getParser(repo repository.Repository) (*fileParser, error) {
 	ns := repo.Config().Namespace
 	if ns == "" {
 		return nil, fmt.Errorf("missing namespace")
@@ -127,7 +128,7 @@ func (s *filesConnector) getParser(repo Repository) (*fileParser, error) {
 	}, nil
 }
 
-func (s *filesConnector) doRead(ctx context.Context, repo Repository, path string, ref string) (*provisioning.ResourceWrapper, error, int) {
+func (s *filesConnector) doRead(ctx context.Context, repo repository.Repository, path string, ref string) (*provisioning.ResourceWrapper, error, int) {
 	info, err := repo.Read(ctx, path, ref)
 	if err != nil {
 		return nil, err, 0
@@ -150,7 +151,7 @@ func (s *filesConnector) doRead(ctx context.Context, repo Repository, path strin
 	return parsed.AsResourceWrapper(), nil, code
 }
 
-func (s *filesConnector) doWrite(ctx context.Context, update bool, repo Repository, path string, ref string, message string, req *http.Request) (runtime.Object, error) {
+func (s *filesConnector) doWrite(ctx context.Context, update bool, repo repository.Repository, path string, ref string, message string, req *http.Request) (runtime.Object, error) {
 	settings := repo.Config().Spec.Editing
 	if update && !settings.Update {
 		return nil, errors.NewForbidden(provisioning.RepositoryResourceInfo.GroupResource(), "updating files not enabled", nil)
@@ -164,7 +165,7 @@ func (s *filesConnector) doWrite(ctx context.Context, update bool, repo Reposito
 		return nil, err
 	}
 
-	info := &FileInfo{
+	info := &repository.FileInfo{
 		Data: data,
 		Path: path,
 		Ref:  ref,
@@ -199,7 +200,7 @@ func (s *filesConnector) doWrite(ctx context.Context, update bool, repo Reposito
 	return parsed.obj, err
 }
 
-func (s *filesConnector) doDelete(ctx context.Context, repo Repository, path string, ref string, message string) (runtime.Object, error) {
+func (s *filesConnector) doDelete(ctx context.Context, repo repository.Repository, path string, ref string, message string) (runtime.Object, error) {
 	settings := repo.Config().Spec.Editing
 	if !settings.Delete {
 		return nil, errors.NewForbidden(provisioning.RepositoryResourceInfo.GroupResource(), "deleting is not supported", nil)
