+++ pkg/registry/apis/provisioning/register.go
@@ -29,6 +29,7 @@ import (
 	provisioning "github.com/grafana/grafana/pkg/apis/provisioning/v0alpha1"
 	grafanaregistry "github.com/grafana/grafana/pkg/apiserver/registry/generic"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/auth"
+	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository"
 	"github.com/grafana/grafana/pkg/services/apiserver/builder"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/grafana/grafana/pkg/setting"
@@ -46,7 +47,7 @@ type ProvisioningAPIBuilder struct {
 
 	features          featuremgmt.FeatureToggles
 	getter            rest.Getter
-	localFileResolver *LocalFolderResolver
+	localFileResolver *repository.LocalFolderResolver
 	logger            *slog.Logger
 	client            *resourceClient
 }
@@ -55,7 +56,7 @@ type ProvisioningAPIBuilder struct {
 // Avoid adding anything that secretly requires additional hidden dependencies
 // like *settings.Cfg or core grafana services that depend on database connections
 func NewProvisioningAPIBuilder(
-	local *LocalFolderResolver,
+	local *repository.LocalFolderResolver,
 	urlProvider func(namespace string) string,
 	webhookSecreteKey string,
 	identities auth.BackgroundIdentityService,
@@ -82,7 +83,7 @@ func RegisterAPIService(
 	if !features.IsEnabledGlobally(featuremgmt.FlagProvisioning) && !features.IsEnabledGlobally(featuremgmt.FlagGrafanaAPIServerWithExperimentalAPIs) {
 		return nil // skip registration unless opting into experimental apis OR the feature specifically
 	}
-	builder := NewProvisioningAPIBuilder(&LocalFolderResolver{
+	builder := NewProvisioningAPIBuilder(&repository.LocalFolderResolver{
 		ProvisioningPath: cfg.ProvisioningPath,
 		DevenvPath:       filepath.Join(cfg.HomePath, "devenv"),
 	}, func(namespace string) string {
@@ -168,7 +169,7 @@ func (b *ProvisioningAPIBuilder) UpdateAPIGroupInfo(apiGroupInfo *genericapiserv
 	return nil
 }
 
-func (b *ProvisioningAPIBuilder) GetRepository(ctx context.Context, name string) (Repository, error) {
+func (b *ProvisioningAPIBuilder) GetRepository(ctx context.Context, name string) (repository.Repository, error) {
 	obj, err := b.getter.Get(ctx, name, &metav1.GetOptions{})
 	if err != nil {
 		return nil, err
@@ -176,7 +177,7 @@ func (b *ProvisioningAPIBuilder) GetRepository(ctx context.Context, name string)
 	return b.asRepository(ctx, obj)
 }
 
-func (b *ProvisioningAPIBuilder) asRepository(ctx context.Context, obj runtime.Object) (Repository, error) {
+func (b *ProvisioningAPIBuilder) asRepository(ctx context.Context, obj runtime.Object) (repository.Repository, error) {
 	if obj == nil {
 		return nil, fmt.Errorf("missing repository object")
 	}
@@ -187,13 +188,13 @@ func (b *ProvisioningAPIBuilder) asRepository(ctx context.Context, obj runtime.O
 
 	switch r.Spec.Type {
 	case provisioning.LocalRepositoryType:
-		return newLocalRepository(r, b.localFileResolver), nil
-	case provisioning.GithubRepositoryType:
-		return newGithubRepository(ctx, r), nil
+		return repository.NewLocal(r, b.localFileResolver), nil
+	case provisioning.GitHubRepositoryType:
+		return repository.NewGitHub(ctx, r), nil
 	case provisioning.S3RepositoryType:
-		return newS3Repository(r), nil
+		return repository.NewS3(r), nil
 	default:
-		return &unknownRepository{config: r}, nil
+		return repository.NewUnknown(r), nil
 	}
 }
 
@@ -346,7 +347,7 @@ func (b *ProvisioningAPIBuilder) Mutate(ctx context.Context, a admission.Attribu
 		return fmt.Errorf("expected repository configuration")
 	}
 
-	if r.Spec.Type == provisioning.GithubRepositoryType {
+	if r.Spec.Type == provisioning.GitHubRepositoryType {
 		if r.Spec.GitHub == nil {
 			return fmt.Errorf("github configuration is required")
 		}
@@ -413,7 +414,7 @@ func (b *ProvisioningAPIBuilder) Validate(ctx context.Context, a admission.Attri
 			cfg.Spec.GitHub, "Local config only valid when type is local"))
 	}
 
-	if cfg.Spec.Type != provisioning.GithubRepositoryType && cfg.Spec.GitHub != nil {
+	if cfg.Spec.Type != provisioning.GitHubRepositoryType && cfg.Spec.GitHub != nil {
 		list = append(list, field.Invalid(field.NewPath("spec", "github"),
 			cfg.Spec.GitHub, "Github config only valid when type is github"))
 	}
